+++ src/view/components/Asset.tsx
@@ -1,9 +1,10 @@
-import BN from "bn.js";
+import BigNumber from "bignumber.js";
 import { FC, useMemo } from "react";
 import styled, { css } from "styled-components";
+import { AppStock } from "../../libs/entries/stock";
 import { ipfsProxy } from "../../libs/service/requestService";
-import { formatAmountValue } from "../../libs/state/decimalsService";
-import { formatCoinValue, useCoinFiat } from "../utils";
+import { formatDecimals } from "../../libs/state/decimalsService";
+import { balanceFormat, fiatFormat } from "../utils";
 import { Gap } from "./Components";
 import { ArrowForwardIcon, BaseLogoIcon } from "./Icons";
 
@@ -13,11 +14,12 @@ export interface AssetProps {
   logoUrl?: string;
   fiat?: string;
   onShow?: () => void;
+  stocks?: AppStock[];
 }
 
 export interface AssetJettonProps extends AssetProps {
-  decimals?: number | string;
-  balance?: string | BN;
+  decimals?: number;
+  balance?: string;
   price?: number;
 }
 
@@ -94,14 +96,22 @@ export const AssetJettonView: FC<AssetJettonProps> = ({
   price,
   ...props
 }) => {
-  const formatted = useMemo(() => {
+  const [formatted, fiat] = useMemo(() => {
     if (!balance) {
-      return "0";
+      return ["0", undefined] as const;
     }
-    return formatCoinValue(formatAmountValue(balance, decimals));
-  }, [balance]);
 
-  const fiat = useCoinFiat(formatted, price);
+    const amount = new BigNumber(balance);
+
+    const formatted = balanceFormat.format(formatDecimals(amount, decimals));
+    let fiat: string | undefined = undefined;
+    if (price) {
+      fiat = fiatFormat.format(
+        formatDecimals(amount.multipliedBy(price), decimals)
+      );
+    }
+    return [formatted, fiat] as const;
+  }, [price, balance]);
 
   return <AssetItemView name={`${formatted} ${name}`} fiat={fiat} {...props} />;
 };
@@ -111,8 +121,13 @@ export const AssetItemView: FC<AssetProps> = ({
   logo,
   logoUrl,
   fiat,
+  stocks,
   onShow,
 }) => {
+  const fiatTitle = useMemo(() => {
+    return stocks && stocks[0] ? stocks[0].dex : undefined;
+  }, [stocks]);
+
   return (
     <Block pointer={onShow != null} onClick={onShow}>
       <ImageBlock>
@@ -133,7 +148,7 @@ export const AssetItemView: FC<AssetProps> = ({
       </ImageBlock>
       <Text>
         <Balance>{name}</Balance>
-        {fiat && <Fiat>{fiat}$</Fiat>}
+        {fiat && <Fiat title={fiatTitle}>{fiat}</Fiat>}
       </Text>
       <Gap />
       {onShow && (