+++ src/view/screen/home/wallet/assets/jetton/send/SendJettonConfirm.tsx
@@ -1,4 +1,3 @@
-import BN from "bn.js";
 import React, { FC, useContext } from "react";
 import { useSearchParams } from "react-router-dom";
 import styled from "styled-components";
@@ -47,7 +46,7 @@ const Quote = styled.div`
 interface ConfirmProps {
   jetton: JettonAsset;
   state: SendJettonState;
-  balance?: BN;
+  balance?: string;
   onSend: (seqNo: number, transactionId?: string) => void;
 }